pyproject.toml
[project]
name = "qwen-vl-object-detect"
version = "0.1.0"
description = "Object Detection with VLM"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "numpy>=2.3.2",
    "openai>=1.99.9",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.55.0",
    "uvicorn>=0.35.0",
]

[tool.uv.sources]
torch = { index = "torch-cpu" }
torchvision = { index = "torch-cpu" }

[[tool.uv.index]]
name = "torch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
